narrow: Explicitly hide recent topics when changing narrow.
User can go from recent topics to stream / topic narrow via
various means, but all go through narrow.active, hence
we make sure all the fixes we do in recent_topics.hide are
actually applied when we hide recent topics and go to
another narrow.

This fixes the bug that narrowing from left sidebar to a stream
takes user to the top of the narrow.